feat: implement new design homepage at cc careers
